# docs/as-contract/openapi-as.yaml
openapi: 3.0.3
info:
  title: RestLink USSD GW — Application Server API
  version: 1.0.0
  description: |
    Dual-mode HTTP AS contract: classic XmlMAPDialog-compatible **text/xml** (default)
    plus greenfield **application/json**. Per-tenant `http_as_wire_format` /
    global `ussd.as.http.wire-format` (blank → XML).

    XML dialog elements, NI path `/ussd`, JSESSIONID, emptyDialogHandshake, and
    AdaptiveTimeout parking: see classic-xml.md.

    Modes:
    - SYNC — pull returns AsResponse with async=false; GW applies CONTINUE/END on live MAP dialog.
    - ASYNC_ACK — pull returns async=true; AS later POSTs /as/callback (or gRPC Callback) while dialog may still be open.
    - BRIDGE — if AS is late past adaptive gate and bridge is armed, GW releases MAP with wait text then NI-pushes on late callback.
    - CLASSIC_NI_SYNC — AS POSTs to /ussd (configurable); multi-turn with JSESSIONID; HTTP response parked until MAP progress (no Thread.sleep).
paths:
  /ussd/pull:
    post:
      summary: Synchronous pull from gateway (MO / continue)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsRequest'
          text/xml:
            schema:
              type: string
              description: Classic XmlMAPDialog-compatible root element dialog (see classic-xml.md)
            example: |
              <?xml version="1.0" encoding="UTF-8"?>
              <dialog type="Begin" appCntx="networkUnstructuredSsContext_version2"
                      networkId="0" localId="corr-1" mapMessagesSize="1">
                <processUnstructuredSSRequest_Request dataCodingScheme="15" string="*100#">
                  <msisdn nai="international_number" npi="ISDN" number="251911000001"/>
                </processUnstructuredSSRequest_Request>
              </dialog>
      responses:
        '200':
          description: Immediate menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsResponse'
            text/xml:
              schema:
                type: string
                description: Classic dialog XML response (CONTINUE/END/ABORT mapped from MAP message children)
  /as/callback:
    post:
      summary: Async callback into gateway (unblock / bridge reconcile)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsResponse'
          text/xml:
            schema:
              type: string
              description: Classic dialog XML late response / ack (dual-mode)
      responses:
        '202':
          description: Accepted
  /ussd:
    post:
      summary: Classic NI sync ingress (default ussd.http.ni-path)
      description: |
        Multi-turn NI dialog. First response may Set-Cookie JSESSIONID;
        subsequent requests must send Cookie. emptyDialogHandshake on dialog
        attributes triggers empty MAP open before payload. Parked responses
        use AdaptiveTimeout — never Thread.sleep.
      requestBody:
        required: true
        content:
          text/xml:
            schema:
              type: string
              description: Classic NI dialog (unstructuredSSRequest_Request / Notify + msisdn)
            example: |
              <?xml version="1.0" encoding="UTF-8"?>
              <dialog mapMessagesSize="1" emptyDialogHandshake="false">
                <unstructuredSSRequest_Request dataCodingScheme="15" string="Press 1 to confirm">
                  <msisdn nai="international_number" npi="ISDN" number="251911000001"/>
                </unstructuredSSRequest_Request>
              </dialog>
          application/json:
            schema:
              $ref: '#/components/schemas/AsRequest'
      responses:
        '200':
          description: NI turn response (may be delayed until MAP dialog advances)
          content:
            text/xml:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/AsResponse'
components:
  schemas:
    AsRequest:
      type: object
      description: |
        GW→AS pull. correlationId is the real push-back key for /as/callback and
        gRPC Callback. sessionId is virtualSessionId (logical). gRPC JSON bytes
        use the same schema.
      properties:
        sessionId: { type: string, description: Logical virtualSessionId }
        correlationId: { type: string, description: Real store/bridge key for late push-back }
        requestId: { type: string }
        generation: { type: integer }
        msisdn: { type: string }
        shortCode: { type: string }
        ussdString: { type: string }
        originatedUssd: { type: string, description: Full UE dialed string (MAP2MAP / MO) }
        codeKind: { type: string, enum: [SHORT, LONG], description: Mark-matched longer dial = LONG }
        networkId: { type: integer }
        virtualBridgeId: { type: string, description: Bridge arm id when armed (usually = correlationId) }
        adaptiveTimeoutMs: { type: integer, format: int64, description: Effective AdaptiveTimeout gate ms }
        asMode: { type: string, enum: [SYNC, ASYNC_ACK, BRIDGE], description: Pull mode hint }
        jsessionId: { type: string, description: Classic NI Cookie JSESSIONID when prior session was gated (HTTP-NI) }
        gateReason: { type: string, description: GATE_EXPIRED | BRIDGED | GATE_NO_BRIDGE when prior session was gated }
        observedEwmaMs: { type: integer, format: int64, description: Observed EWMA latency ms for networkId when known }
    AsGatedNotify:
      type: object
      description: |
        GW→AS classic XML (preferred) or JSON body POSTed to short-code asUrl when AdaptiveTimeout/bridge gate fires.
        XML encoder ClassicDialogXmlCodec.encodeGatedPush. HTTP client session id gated-{correlationId}.
      properties:
        correlationId: { type: string }
        sessionId: { type: string }
        virtualBridgeId: { type: string }
        jsessionId: { type: string }
        adaptiveTimeoutMs: { type: integer, format: int64 }
        observedEwmaMs: { type: integer, format: int64 }
        gateReason: { type: string }
        action: { type: string, enum: [ABORT] }
      type: object
      description: |
        AS→GW pull response or callback. Echo correlationId (required for push-back);
        virtualBridgeId / sessionId accepted as fallbacks if correlationId omitted.
      properties:
        correlationId: { type: string, description: Real store/bridge key }
        requestId: { type: string }
        generation: { type: integer }
        text: { type: string }
        action: { type: string, enum: [CONTINUE, END, ABORT] }
        async: { type: boolean, description: If true, AS will POST /as/callback or gRPC Callback later }
        alphabet: { type: string, enum: [ucs7, ucs8, unicode, auto] }
        sessionId: { type: string, description: Optional echo of virtualSessionId }
        virtualBridgeId: { type: string, description: Optional echo for late resolve }
        adaptiveTimeoutMs: { type: integer, format: int64, description: Optional gate echo }